Move the edit project team code
This conceptually relates to the edit assigned to and added by code that
is presented to users under 'internal contacts'.

This works moves it into the same namespace and tidies as we go. We've
collected all the feature specs into one place and deleted duplicate
specs.
