+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   uses a select, instead an email address must be entered.
 - The autocomplete for editing user who added a project no longer loads all
   users onto the page, instead it fetches the users as they are typed.
+- The path for changing a project team has changed along with the layout of the
+  form for consistency.
 
 ### Fixed
