+++ app/views/internal_contacts/projects/edit_team.html.erb
@@ -1,14 +1,14 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= form_for @project, url: project_assign_team_path(@project), method: :post do |form| %>
+    <%= form_for @project, url: project_internal_contacts_team_path(@project), method: :put do |form| %>
       <%= form.govuk_error_summary %>
 
-      <span class="govuk-caption-m">URN <%= @project.urn %></span>
+      <span class="govuk-caption-l">URN <%= @project.urn %></span>
       <%= form.govuk_collection_radio_buttons :team,
             Project.teams.keys,
             :to_s,
             ->(team) { I18n.t("teams.#{team}") },
-            legend: {text: t("assignment.assign_team.title", school_name: @project.establishment.name)} %>
+            legend: {text: t("assignment.assign_team.title", school_name: @project.establishment.name), size: "xl", tag: :h1} %>
       <%= form.govuk_submit do %>
         <%= govuk_link_to t("cancel"), project_internal_contacts_path(@project) %>
       <% end %>
