+++ config/routes.rb
@@ -27,14 +27,14 @@ Rails.application.routes.draw do
     namespace :assign, controller: "/assignments" do
       get "team-lead", action: :assign_team_leader
       post "team-lead", action: :update_team_leader
-      get "team", action: :assign_team
-      post "team", action: :update_team
     end
     namespace :internal_contacts, path: "internal-contacts" do
       get "assigned-user/edit", action: :edit_assigned_user
       put "assigned-user/", action: :update_assigned_user
       get "added-by-user/edit", action: :edit_added_by_user
       put "added-by-user/", action: :update_added_by_user
+      get "team/edit", action: :edit_team
+      put "team", action: :update_team
     end
   end
