+++ spec/features/internal_contacts/users_can_edit_the_assigned_user_spec.rb
